Rename `base`.
